requirements.txt
fastapi
pydantic
SQLAlchemy
bcrypt
uvicorn
black
python-jose
python-dotenv
psycopg2-binary